chatbot-rag/requirements.txt
# ─── Core ───────────────────────────────────────────────
langchain>=0.3.17
langchain-community>=0.3.16
langchain-openai>=0.3.4
langchain-chroma>=0.2.2
openai>=1.61.0

# ─── Vector Database ────────────────────────────────────
chromadb>=0.6.3

# ─── Document Processing ────────────────────────────────
pypdf>=5.0.0
docx2txt>=0.8

# ─── Web Interfaces ─────────────────────────────────────
streamlit>=1.42.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
flask>=3.0.0
twilio>=9.0.0

# ─── Utilities ──────────────────────────────────────────
python-dotenv>=1.0.0
tiktoken>=0.8.0
python-multipart>=0.0.20
pydantic>=2.10.0
